Move tests to be in the right packages and rename UriSupplier
